requirements.txt
flask==2.3.3
flask-cors==4.0.0
spacy==3.7.2
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.0/en_core_web_sm-3.7.0.tar.gz#egg=en_core_web_sm
pdfplumber==0.10.2
python-docx==0.8.11
scikit-learn==1.3.2
transformers==4.35.0
torch==2.1.0
sentence-transformers==2.2.2
pandas==2.1.3
numpy==1.24.3
plotly==5.18.0
matplotlib==3.8.2
seaborn==0.13.0
python-dotenv==1.0.0
fuzzywuzzy==0.18.0
python-Levenshtein==0.25.1